[sensorfw] Fix MER#983 revert accel x & y values.

+++ filters/avgaccfilter/avgaccfilter.h
@@ -53,9 +53,9 @@ private:
     unsigned int avgBufferSize;
     qreal filterFactor;
 
-    int averageX;
-    int averageY;
-    int averageZ;
+    qreal averageX;
+    qreal averageY;
+    qreal averageZ;
 
     QList<TimedXyzData> avgAccelBuffer;
